mlir: extract MlirCpuTarget base for the shared CPU codegen driver
